Implement JWT based user authentication
This commit introduces JSON Web Token (JWT) based user authentication in the application. It includes JWT decoding in the UserSlice to enable user login and logout actions and the addition of user state with access and refresh tokens. Corresponding adjustments have been made in the login and register components as well, with enhanced error handling and redirection features.